Extract table pagination component.
