+++ react-components/src/util/table/messages.js
@@ -1,0 +1,9 @@
+const intlData = {
+    "locales": "en-US",
+    "messages": {
+        "firstPage": "First Page",
+        "lastPage": "Last Page",
+        "nextPage": "Next Page",
+        "prevPage": "Previous Page",
+    }
+}
